+++ web/src/App.css
@@ -184,7 +184,7 @@ body {
 
 body {
   margin: 0px 0px 0px 0px;
-  padding: 80px 0px 0px 0px;
+  padding: 70px 0px 0px 0px;
   color: #555;
   display: flex;
   flex-direction: column;
